Mejoras generales de la aplicacion mas juego Se ha dado la opcion de cancelar reserva al usuario y ademas si no hay mesas disponibles para la fecha y horario, se le informa automaticamente. Se ha aniadido la funcion de logout en todas las vistas. Ademas se a adelantado un buen porcentaje del juego por un postre gratis.

+++ src/pages/alta-duenio-supervisor/alta-duenio-supervisor.ts
@@ -19,6 +19,7 @@ export class AltaDuenioSupervisorPage {
 
   public firebase = firebase;
   public db = firebase.firestore();
+  public usuario: any;
 
   public correo: string;
   public clave: string;
@@ -45,7 +46,7 @@ export class AltaDuenioSupervisorPage {
     private camera: Camera,
     private barcodeScanner: BarcodeScanner) {
 
-    this.authInstance.auth.signInWithEmailAndPassword("example@gmail.com", "123456");
+    this.usuario = JSON.parse(localStorage.getItem("usuario"));
   }
 
   ionViewDidLoad() {
@@ -187,7 +188,7 @@ export class AltaDuenioSupervisorPage {
 
   InicializarLectorQR() {
 
-    let options = { prompt : "Escaneá el DNI", formats: "PDF_417" };
+    let options = { prompt: "Escaneá el DNI", formats: "PDF_417" };
 
     this.barcodeScanner.scan(options).then(barcodeData => {
 
@@ -249,8 +250,42 @@ export class AltaDuenioSupervisorPage {
 
   Logout() {
 
-    localStorage.clear();
-    this.navCtrl.setRoot(LoginPage);
+    let usuariosRef = this.firebase.database().ref("usuarios");
+
+    usuariosRef.once("value", (snap) => {
+
+      let data = snap.val();
+
+      for (let item in data) {
+
+        if (data[item].correo == this.usuario.correo) {
+
+          usuariosRef.child(item).update({
+            logueado: false
+          }).then(() => {
+            if (this.usuario.tipo == "mozo"
+              || this.usuario.tipo == "cocinero"
+              || this.usuario.tipo == "bartender"
+              || this.usuario.tipo == "metre"
+              || this.usuario.tipo == "repartidor") {
+
+              // Para redireccionar a la encuesta de axel.
+              // localStorage.setItem("desloguear", "true");
+              // this.navCtrl.setRoot(EncuestaDeEmpleadoPage);
+
+              localStorage.clear();
+              this.navCtrl.setRoot(LoginPage);
+            } else {
+
+              localStorage.clear();
+              this.navCtrl.setRoot(LoginPage);
+            }
+          });
+
+          break;
+        }
+      }
+    });
   }
 
 }
